requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
requests==2.32.2
Pillow==10.3.0
lxml==5.2.1
selenium==4.20.0
webdriver-manager==4.0.2